gp-import: supported double format of tempo for gp8

+++ src/importexport/guitarpro/internal/gtp/gp67dombuilder.cpp
@@ -283,7 +283,7 @@ std::vector<GPMasterTracks::Automation> GP67DomBuilder::readTempoMap(XmlDomNode*
                 tempo.type = GPMasterTracks::Automation::Type::tempo;
                 String str = currentAutomation.firstChildElement("Value").toElement().text();
                 StringList tempoValue = str.split(u' ');
-                tempo.value = static_cast<int>(tempoValue[0].toDouble());
+                tempo.value = tempoValue[0].toDouble();
                 tempo.tempoUnit = tempoValue.size() > 1 ? static_cast<int>(tempoValue[1].toDouble()) : 0;
                 tempo.bar = currentAutomation.firstChildElement("Bar").text().toInt();
                 tempo.position = currentAutomation.firstChildElement("Position").text().toFloat();
